requirements.txt
Django
ipdb
pytest-django
